Moved show_anatomical_slices to viz.utils
